% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_indexing.R
\name{add_sample_info}
\alias{add_sample_info}
\title{Add Sample Info Based on Well's Location}
\usage{
add_sample_info(sce, fryDir, bc_filter)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object produced from
\code{\link[DPLabR]{load_split_seq}} object}

\item{fryDir}{An Alevin-Fry main directory}

\item{bc_filter}{Logical to assess whether the sample file has been filterd or not}
}
\value{
A A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
This function add the sample info from a .txt. This file is obtained running a custom python
script from the SPLiT-seq pipeline. Info is added to each cell based on the original location
in the multiwell
}
