% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{Annot_DE}
\alias{Annot_DE}
\title{Annotate DE Table}
\usage{
Annot_DE(df, log2FC = 1, padjust = 0.1)
}
\arguments{
\item{df}{A Dataframe obtained from a differential expression test like \code{DESeq2::DESeq()}.}

\item{log2FC}{A numeric value specifying log2FoldChange threshold}

\item{padjust}{A numeric value specifying padjusted threshold}
}
\value{
A Dataframe annotated with Upregulated and Downregulated genes. Thresholds are used defined
}
\description{
Useful function tofilter the excel file or to do a volcano plot coloring the DE and non DE genes
requires a gene expression table with log2FC and padjust column for each gene,
like the one obtained from \code{DESeq()} call.
}
