% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_enrichments.R
\name{fract_to_dec}
\alias{fract_to_dec}
\title{Fraction to Number}
\usage{
fract_to_dec(x)
}
\arguments{
\item{x}{fraction from GO enrichment to transform to a float}
}
\value{
a float
}
\description{
util function to transform a fraction to a number from the output of GO enrichment
}
