% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_geom_boxplot.R
\name{geom_boxplot_noOut}
\alias{geom_boxplot_noOut}
\title{Custom geom_boxplot}
\usage{
geom_boxplot_noOut(size = 0.8, width = 0.8, position = "identity")
}
\arguments{
\item{size}{An integer controlling the size of the boxes. Default is 0.8}

\item{width}{An integer controlling the width of the boxes. Default is 0.8}

\item{position}{A character specifiyng the position. Default is "identity"}
}
\value{
A geom layer
}
\description{
A Custom geom_boxplot for a \code{\link[ggplot2]{ggplot}} object. It has been computed to remove outliers
if too extreme and make a plot unreadable
}
