% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage_custom.R
\name{get_coverage_custom}
\alias{get_coverage_custom}
\title{Quantify ChIPseq Scores}
\usage{
get_coverage_custom(bw_list, regions, operation)
}
\arguments{
\item{bw_list}{A named list of .bw files}

\item{regions}{A named list of genomic regions in BED file}

\item{operation}{Operation to compute from \code{\link[megadepth]{get_coverage}} function. Defaul = "mean"}
}
\value{
A tibble with quantifications
}
\description{
Custom function to quantify ChIPseq scores in multiple .bw files
across multiple genomics regions in BED format
}
