% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_log_info.R
\name{get_log_multimapped}
\alias{get_log_multimapped}
\title{Function to parse aligment log file}
\usage{
get_log_multimapped(x)
}
\arguments{
\item{x}{A named List of character vectors with one element for each line}
}
\value{
a Named list with one numerical value per element of the list
}
\description{
This is a function to parse log information from alignment files. For now it retrieves only
discarded multimapped reads. Needs to be expanded in functionalities.
}
