% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_enrichments.R
\name{goEnrichment}
\alias{goEnrichment}
\title{Custom Go Function}
\usage{
goEnrichment(df, ont = "BP", db = org.Mm.eg.db)
}
\arguments{
\item{df}{A 2 column dataframe, one called \code{ENTREZID} and the other one \code{SYMBOL} representing genes to be tested.}

\item{ont}{The ontology to use. Default is \code{BP}}

\item{db}{A character specifying with specie database to use. Default is Mouse but will be override by an argument in the
main function \code{\link[DPLabR]{enrichments}}}
}
\value{
A dataframe with ontologies filterd by p.adjusted < 0.05
}
\description{
This is the custom function run by the previous analysis. It performs GO analysis correcting with \code{BH}
}
