% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_enrichments.R
\name{KEGGenrichment}
\alias{KEGGenrichment}
\title{Custom KEGG function}
\usage{
KEGGenrichment(df, org = "mmu", db = org.Mm.eg.db)
}
\arguments{
\item{df}{A 2 column dataframe, one called \code{ENTREZID} and the other one \code{SYMBOL} representing genes to be tested.}

\item{org}{A character specifying which specie to use. Default "mmu" but is overridden in the main function.}

\item{db}{A character specifying with specie database to use. Default is Mouse but will be override by an argument in the
main function \code{\link[DPLabR]{enrichments}}}
}
\value{
A dataframe with KEGG pathways filterd by p.adjusted < 0.05
}
\description{
This is the custom function run by the previous analysis. It performs KEGG analysis correcting with \code{BH}
}
