% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorial_indexing.R
\name{load_split_seq}
\alias{load_split_seq}
\title{Load and convert ensembl genes to gene Symbol}
\usage{
load_split_seq(fryDir, outputFormat, expType)
}
\arguments{
\item{fryDir}{An Alevin-Fry main directory. It breaks as of now if a custom alevin
directory have been specified}

\item{outputFormat}{One of the available outputformat of the \code{\link[fishpond]{loadFry}}}

\item{expType}{One of the available experimental types between Mixed, Human or Mouse}
}
\value{
A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
This function loads in the genes x BCs matrix and creates a
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
