% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_enrichments.R
\name{msig_db_enrichment}
\alias{msig_db_enrichment}
\title{Custom Gene Set Enrichments}
\usage{
msig_db_enrichment(df, pathways.gmt)
}
\arguments{
\item{df}{A 2 column dataframe, one called \code{ENTREZID} and the other one \code{SYMBOL} representing genes to be tested.}

\item{pathways.gmt}{A dataframe containing pathways/gene sets to be tested}
}
\value{
A dataframe with gene set analysis filtered by p.adjusted < 0.5
}
\description{
This is the custom function run by the previous analysis. It performs gene set enrichment analysis.
}
