% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_enrichments.R
\name{PAenrichment}
\alias{PAenrichment}
\title{Custom Pathway function}
\usage{
PAenrichment(df, org = "mouse")
}
\arguments{
\item{df}{A 2 column dataframe, one called \code{ENTREZID} and the other one \code{SYMBOL} representing genes to be tested.}

\item{org}{A character specifying with specie organism to use. Default is Mouse but will be override by an argument in the
main function \code{\link[DPLabR]{enrichments}}}
}
\value{
A dataframe with Pathways analysis filtered by p.adjusted < 0.5
}
\description{
This is the custom function run by the previous analysis. It performs Pathway analysis.
}
