% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_enrichments.R
\name{plot_fgsea}
\alias{plot_fgsea}
\title{Custom GSEA Plot}
\usage{
plot_fgsea(df, specie = "human", query, NES_cutoff = 1)
}
\arguments{
\item{df}{A dataframe containing the results of differential expression analysis. It assumes the structure we give to those dataframes in our pipelines
and therefore should have column \code{Geneid} at index 1 and \code{log2FoldChange} at index 3.}

\item{specie}{A character. What specie should be used. Default is "human". Anything different than that is evaluated as "mouse"}

\item{query}{A character. What pathway should be tested. By default every gene set that partially matches that pathway. (e.g "WNT", "APOPTOSIS")}

\item{NES_cutoff}{A numeric. What Normalized Enrichment Score to use as a cutoff (bidirectional). It doesn't assume any significance at this point.}
}
\value{
A list with 4 elements for convenience.
}
\description{
This function creates a custom GSEA plot in the form a volcano where all the genesets tested are reported, provided that they
pass a \code{NES} threshold
}
