% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Volcano Plot}
\usage{
VolcanoPlot(
  df,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  labelSize = 7,
  pval = 0.1,
  log2FC = 1,
  raw_numbers = TRUE
)
}
\arguments{
\item{df}{A dataframe containing differentally expressed genes as an output from \code{\link[DESeq2]{DESeq}}.
For the DP lab, this is the output of the contrast section of the pipeline.}

\item{xlim}{User defined xlim}

\item{ylim}{User defined ylim}

\item{main}{User defined main title}

\item{labelSize}{labelSize of the number or percentage of Upregulated or Downregulated genes}

\item{pval}{Statistical threshold to draw pvalue significance line. Default is 0.1}

\item{log2FC}{Statistical threshold to draw log3FoldChange line. Default is 0.1}

\item{raw_numbers}{Logical. Whether to draw the number of UP/DOWN genes as an integer or a fraction}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Custom code that creates a \code{\link[ggplot2]{ggplot}} object plotting differentially
expressed genes.
}
