% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{VolcanoPlot_Genes}
\alias{VolcanoPlot_Genes}
\title{Volcano Plot with Genes Info}
\usage{
VolcanoPlot_Genes(
  df,
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  labelSize = 7,
  pval = 0.1,
  log2FC = 1,
  Genes = NULL
)
}
\arguments{
\item{df}{A  dataframe containing differentally expressed genes as an output from \code{\link[DESeq2]{DESeq}}.
For the DP lab, this is the output of the contrast section of the pipeline.}

\item{xlim}{User defined xlim}

\item{ylim}{User defined ylim}

\item{main}{User defined main title}

\item{labelSize}{labelSize of the number or percentage of Upregulated or Downregulated genes}

\item{pval}{Statistical threshold to draw pvalue significance line. Default is 0.1}

\item{log2FC}{Statistical threshold to draw log3FoldChange line. Default is 0.1}

\item{Genes}{A user defined list of genes of interests to be highlighted}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Creates a \code{\link[ggplot2]{ggplot}} object plotting differentially
expressed genes and highlighting a user defined list of genes
}
